refactoring editor
