+++ bot/handlers/editor/__init__.py
@@ -1,0 +1,1 @@
+from bot.handlers.editor import hide, choose, returning, place
